refactors kill methods into action.py
